+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fixes incorrect generation of world ID codes when meetings were selected for unpublish on virtual-na.org (thanks @anitak!)
 - Publish on virtualna.org selector wasn't appearing on quickedit. It is now :)
 - Deprecated support for BMLT 2.x
+- We now support retrieving/editing BMLT meeting comments in the quickedit dialog (thanks @brustar64 !)
 
 ## 1.1.17 (Oct 14, 2024)
 - Fixes contact phone number format limitation on end user form (thanks @adamH !)